regression fixes react-env
